+++ resources/js/components/ui/text-field.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import * as React from 'react';
+import { useState } from 'react';
 
 import type { TextInputDOMProps } from '@react-types/shared';
 import { IconEye, IconEyeClosed } from 'justd-icons';
@@ -9,11 +9,12 @@ import {
   TextField as TextFieldPrimitive,
   type TextFieldProps as TextFieldPrimitiveProps
 } from 'react-aria-components';
+import { twJoin } from 'tailwind-merge';
 
 import type { FieldProps } from './field';
-import { Description, FieldError, FieldGroup, fieldGroupPrefixStyles, Input, Label } from './field';
+import { Description, FieldError, FieldGroup, Input, Label } from './field';
 import { Loader } from './loader';
-import { ctr } from './primitive';
+import { composeTailwindRenderProps } from './primitive';
 
 type InputType = Exclude<TextInputDOMProps['type'], 'password'>;
 
@@ -49,16 +50,28 @@ const TextField = ({
   type,
   ...props
 }: TextFieldProps) => {
-  const [isPasswordVisible, setIsPasswordVisible] = React.useState(false);
+  const [isPasswordVisible, setIsPasswordVisible] = useState(false);
   const inputType = isRevealable ? (isPasswordVisible ? 'text' : 'password') : type;
-
   const handleTogglePasswordVisibility = () => {
     setIsPasswordVisible((prev) => !prev);
   };
   return (
-    <TextFieldPrimitive type={inputType} {...props} className={ctr(className, 'group flex flex-col gap-y-1.5')}>
+    <TextFieldPrimitive
+      type={inputType}
+      {...props}
+      className={composeTailwindRenderProps(className, 'group flex flex-col gap-y-1.5')}
+    >
       {label && <Label>{label}</Label>}
-      <FieldGroup data-loading={isPending ? 'true' : undefined} className={fieldGroupPrefixStyles({ className })}>
+      <FieldGroup
+        isInvalid={!!errorMessage}
+        isDisabled={props.isDisabled}
+        className={twJoin(
+          '**:[button]:size-7 **:[button]:shrink-0 **:[button]:p-0',
+          '[&>[data-slot=suffix]>button]:mr-[calc(var(--spacing)*-1.15)] [&>[data-slot=suffix]>button]:rounded-md [&>[data-slot=suffix]>button]:data-focus-visible:outline-1 [&>[data-slot=suffix]>button]:data-focus-visible:outline-offset-1',
+          '[&>[data-slot=prefix]>button]:mr-[calc(var(--spacing)*-1.15)] [&>[data-slot=prefix]>button]:rounded-md [&>[data-slot=prefix]>button]:data-focus-visible:outline-1 [&>[data-slot=prefix]>button]:data-focus-visible:outline-offset-1'
+        )}
+        data-loading={isPending ? 'true' : undefined}
+      >
         {prefix ? (
           <span data-slot="prefix" className="atrs x2e2">
             {prefix}
@@ -70,16 +83,14 @@ const TextField = ({
             type="button"
             aria-label="Toggle password visibility"
             onPress={handleTogglePasswordVisibility}
-            className="atrs relative [&>[data-slot=icon]]:text-muted-fg focus:outline-none focus-visible:ring-1 focus-visible:ring-primary rounded"
+            className="relative mr-1 grid shrink-0 place-content-center rounded-sm border-transparent outline-hidden data-focus-visible:*:data-[slot=icon]:text-primary *:data-[slot=icon]:text-muted-fg"
           >
-            <>{isPasswordVisible ? <IconEyeClosed /> : <IconEye />}</>
+            {isPasswordVisible ? <IconEyeClosed /> : <IconEye />}
           </ButtonPrimitive>
         ) : isPending ? (
           <Loader variant="spin" data-slot="suffix" />
         ) : suffix ? (
-          <span data-slot="suffix" className="atrs x2e2">
-            {suffix}
-          </span>
+          <span data-slot="suffix">{suffix}</span>
         ) : null}
       </FieldGroup>
       {description && <Description>{description}</Description>}